Allow for service tags to have colons in them
Currently the splitting code selects only the first part of the
service tag if the service tag contained a colon. For example given the
tag `service:foo:bar`, the forwarder would select the service name of
`foo` and not `foo:bar`.

This change modifies the splitting behaviour to correctly return
`foo:bar` in this case.